refactor: apply spotless
