+++ core/src/main/java/io/naryo/application/configuration/source/model/node/subscription/factory/DefaultBlockSubscriptionFactory.java
@@ -1,4 +1,9 @@
-package io.naryo.application.configuration.factory;
+package io.naryo.application.configuration.source.model.node.subscription.factory;
+
+import java.math.BigInteger;
+import java.util.Optional;
+import java.util.function.Consumer;
+import java.util.function.Supplier;
 
 import io.naryo.application.configuration.source.model.node.subscription.BlockSubscriptionDescriptor;
 import io.naryo.application.configuration.source.model.node.subscription.PollBlockSubscriptionDescriptor;
@@ -8,12 +13,7 @@ import io.naryo.domain.node.subscription.block.method.poll.Interval;
 import io.naryo.domain.node.subscription.block.method.poll.PollBlockSubscriptionMethodConfiguration;
 import io.naryo.domain.node.subscription.block.method.pubsub.PubSubBlockSubscriptionMethodConfiguration;
 
-import java.math.BigInteger;
-import java.util.Optional;
-import java.util.function.Consumer;
-import java.util.function.Supplier;
-
-public final class DefaultBlockSubscriptionFactory implements BlockSubscriptionFactory{
+public final class DefaultBlockSubscriptionFactory implements BlockSubscriptionFactory {
 
     public static final BigInteger INITIAL_BLOCK = BigInteger.valueOf(-1);
     public static final BigInteger CONFIRMATION_BLOCKS = BigInteger.valueOf(12);
@@ -27,37 +27,41 @@ public final class DefaultBlockSubscriptionFactory implements BlockSubscriptionF
 
         this.applyDefaults(descriptor);
 
-        var method = switch (descriptor.getMethod()) {
-            case POLL -> {
-                var poll = (PollBlockSubscriptionDescriptor) descriptor;
-                var interval = poll.getInterval().orElse(null);
-                yield new PollBlockSubscriptionMethodConfiguration(new Interval(interval));
-            }
-            case PUBSUB -> new PubSubBlockSubscriptionMethodConfiguration();
-        };
+        var method =
+                switch (descriptor.getMethod()) {
+                    case POLL -> {
+                        var poll = (PollBlockSubscriptionDescriptor) descriptor;
+                        var interval = poll.getInterval().orElse(null);
+                        yield new PollBlockSubscriptionMethodConfiguration(new Interval(interval));
+                    }
+                    case PUBSUB -> new PubSubBlockSubscriptionMethodConfiguration();
+                };
 
         return new BlockSubscriptionConfiguration(
-            method,
-            descriptor.getInitialBlock().orElseThrow(),
-            new NonNegativeBlockNumber(descriptor.getConfirmationBlocks().orElseThrow()),
-            new NonNegativeBlockNumber(descriptor.getMissingTxRetryBlocks().orElseThrow()),
-            new NonNegativeBlockNumber(descriptor.getEventInvalidationBlockThreshold().orElseThrow()),
-            new NonNegativeBlockNumber(descriptor.getReplayBlockOffset().orElseThrow()),
-            new NonNegativeBlockNumber(descriptor.getSyncBlockLimit().orElseThrow()));
+                method,
+                descriptor.getInitialBlock().orElseThrow(),
+                new NonNegativeBlockNumber(descriptor.getConfirmationBlocks().orElseThrow()),
+                new NonNegativeBlockNumber(descriptor.getMissingTxRetryBlocks().orElseThrow()),
+                new NonNegativeBlockNumber(
+                        descriptor.getEventInvalidationBlockThreshold().orElseThrow()),
+                new NonNegativeBlockNumber(descriptor.getReplayBlockOffset().orElseThrow()),
+                new NonNegativeBlockNumber(descriptor.getSyncBlockLimit().orElseThrow()));
     }
 
     public void applyDefaults(BlockSubscriptionDescriptor d) {
         setDefault(d::getInitialBlock, d::setInitialBlock, INITIAL_BLOCK);
         setDefault(d::getConfirmationBlocks, d::setConfirmationBlocks, CONFIRMATION_BLOCKS);
         setDefault(d::getMissingTxRetryBlocks, d::setMissingTxRetryBlocks, MISSING_TX_RETRY_BLOCKS);
-        setDefault(d::getEventInvalidationBlockThreshold, d::setEventInvalidationBlockThreshold,
-            EVENT_INVALIDATION_BLOCK_THRESHOLD);
+        setDefault(
+                d::getEventInvalidationBlockThreshold,
+                d::setEventInvalidationBlockThreshold,
+                EVENT_INVALIDATION_BLOCK_THRESHOLD);
         setDefault(d::getReplayBlockOffset, d::setReplayBlockOffset, REPLAY_BLOCK_OFFSET);
         setDefault(d::getSyncBlockLimit, d::setSyncBlockLimit, SYNC_BLOCK_LIMIT);
     }
 
     private static <T> void setDefault(
-        Supplier<Optional<T>> getter, Consumer<T> setter, T defaultValue) {
+            Supplier<Optional<T>> getter, Consumer<T> setter, T defaultValue) {
         setter.accept(getter.get().orElse(defaultValue));
     }
 }
